src: remove unsound async handle manipulation
The uv_async_t handle is never initialized so it's not safe to ref
or unref it.  It's surprising the module doesn't crash more often.

The purported reason is to keep the event loop alive when there is
in-flight work but uv_queue_work() does that anyway so there really
should be no reason to have it around.